Temporary binary buffer exception
It isn't good to mix strings with raw binary data. We should check if the data is binary or string. We can check this two ways, either we convert all messages to bytearray, or we try to decode the message and use an except to notify us when the data is raw binary code.

- Allow binary data to exceed the number 127